Agrego entidades

+++ Dominio/Modelos/IngredienteModelo.cs
@@ -10,7 +10,7 @@ namespace Dominio.Modelos
     {
         public int IdIngrediente {  get; set; }
         public string Nombre { get; set; }
-        public float Cantidad { get; set; }
+        public double Cantidad { get; set; }
         public UnidadMedidaModelo Unidad { get; set; }
         public decimal Costo { get; set; }
         public string Proveedor { get; set; }
